refactor(frontend): Rename ICRC components for adding tokens
