+++ src/frontend/src/icp/components/tokens/IcAddIcrcTokenForm.svelte
