+++ src/frontend/src/icp/components/tokens/IcAddIcrcTokenReview.svelte
