+++ src/frontend/src/lib/components/tokens/TokenModal.svelte
@@ -5,7 +5,7 @@
 	import type { Snippet } from 'svelte';
 	import { erc20UserTokensStore } from '$eth/stores/erc20-user-tokens.store';
 	import { isTokenErc20UserToken } from '$eth/utils/erc20.utils';
-	import IcAddTokenForm from '$icp/components/tokens/IcAddTokenForm.svelte';
+	import IcAddIcrcTokenForm from '$icp/components/tokens/IcAddIcrcTokenForm.svelte';
 	import { assertIndexLedgerId } from '$icp/services/ic-add-custom-tokens.service';
 	import { loadCustomTokens } from '$icp/services/icrc.services';
 	import { icrcCustomTokensStore } from '$icp/stores/icrc-custom-tokens.store';
@@ -363,7 +363,7 @@
 					networkName={token.network.name}
 				/>
 
-				<IcAddTokenForm
+				<IcAddIcrcTokenForm
 					editMode
 					ledgerCanisterId={token.ledgerCanisterId}
 					bind:indexCanisterId={icrcTokenIndexCanisterId}
